Split LibraryDreamService into multiple files and fix various bugs
