+++ app/src/main/java/org/jellyfin/androidtv/integration/dream/composable/DreamContentLogo.kt
@@ -1,0 +1,26 @@
+package org.jellyfin.androidtv.integration.dream.composable
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import org.jellyfin.androidtv.R
+
+@Composable
+fun DreamContentLogo() = Box(
+	modifier = Modifier.fillMaxSize(),
+) {
+	Image(
+		painter = painterResource(R.drawable.app_logo),
+		contentDescription = stringResource(R.string.app_name),
+		modifier = Modifier
+			.align(Alignment.Center)
+			.width(400.dp)
+	)
+}
